refactor: add dropdown hook
